Create general ssml functions and move UserData.
