--- functions/src/responses/errorResponse.ts
@@ -1,3 +1,0 @@
-const errorResponse = 'Something went horribly wrong';
-
-export { errorResponse };
